refactor: organize shared frontend utilities
