+++ utils/measure/frontend/src/components/settings/view.ts
@@ -3,8 +3,8 @@ import { customElement, property, state } from "lit/decorators.js";
 import { createRef, ref } from "lit/directives/ref.js";
 import type { AppSettings, AppSettingsUpdate, Capabilities, ContributionAuthDeviceStatus, ContributionAuthState, ContributionDeviceFlow, EntityDescriptor, MeasureParameterName, PowerMeterDiagnostic, SettingsSection, ShellyDiscoveryDevice } from "../../types";
 import { DEFAULT_SHELLY_USERNAME, settingsFromForm } from "../../power-meter";
-import { formRaw, formText, formTextOrNull } from "../../form";
-import { emit } from "../../events";
+import { emit } from "../../utils/events";
+import { formRaw, formText, formTextOrNull } from "../../utils/form";
 import { sharedStyles } from "../../styles";
 import "./github-section";
 import "./power-meter-section";
