+++ utils/measure/frontend/src/components/setup/tuning-section.ts
@@ -1,7 +1,7 @@
 import { LitElement, html, nothing } from "lit";
 import { customElement, property } from "lit/decorators.js";
 import type { Capabilities, MeasureDefinition, MeasureParameter, MeasureParameterName, MeasurementRequest } from "../../types";
-import { emit } from "../../events";
+import { emit } from "../../utils/events";
 import { gatedParameters } from "../../measure-definition";
 import { numberField } from "../shared/fields";
