+++ utils/measure/frontend/src/review-summary.ts
@@ -1,4 +1,4 @@
-import { duration as formatDuration } from "./format";
+import { duration as formatDuration } from "./utils/format";
 import { requestFieldValue } from "./measure-definition";
 import { summarize } from "./power-meter";
 import type { FormField, MeasureDefinition, MeasurementRequest, PreflightResponse } from "./types";
