+++ utils/measure/frontend/src/utils/format.ts
